+++ Core/function.h
@@ -85,7 +85,7 @@ namespace X
 			}
 			virtual void GetBaseScopes(std::vector<AST::Scope*>& bases) override;
 			virtual int QueryMethod(const char* name, int* pFlags) override;
-			virtual bool GetIndexValue(int idx, Value& v) override;
+			virtual bool GetIndexValue(long long idx, Value& v) override;
 
 			std::string GetDoc()
 			{
