+++ Core/moduleobject.h
@@ -59,7 +59,7 @@ namespace X
 				return GetABIString(str);
 			}
 			virtual int QueryMethod(const char* name, int* pFlags = nullptr) override;
-			virtual bool GetIndexValue(int idx, Value& v) override;
+			virtual bool GetIndexValue(long long idx, Value& v) override;
 			//return size of Module's member count
 			virtual long long Size() override;
 		};
